move tests mod to its own file
